# obai_fullstack_package-v2/backend/src/main/java/com/obai/platform/controller/admin/AdminGroupCustomerLeadController.java
package com.obai.platform.controller.admin;

import com.obai.platform.audit.OperationAudit;
import com.obai.platform.common.ApiResponse;
import com.obai.platform.common.EntityPatchUtil;
import com.obai.platform.entity.GroupCustomerLead;
import com.obai.platform.repository.GroupCustomerLeadRepository;
import java.util.List;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/admin/group-leads")
public class AdminGroupCustomerLeadController {
    private final GroupCustomerLeadRepository repository;

    public AdminGroupCustomerLeadController(GroupCustomerLeadRepository repository) {
        this.repository = repository;
    }

    @GetMapping
    public ApiResponse<List<GroupCustomerLead>> list() {
        return ApiResponse.ok(repository.findAll());
    }

    @GetMapping("/{id}")
    public ApiResponse<GroupCustomerLead> detail(@PathVariable Long id) {
        return ApiResponse.ok(repository.findById(id).orElseThrow());
    }

    @PostMapping
    @OperationAudit(module = "团体客户", action = "新增")
    public ApiResponse<GroupCustomerLead> create(@RequestBody GroupCustomerLead body) {
        return ApiResponse.ok(repository.save(body));
    }

    @PutMapping("/{id}")
    @OperationAudit(module = "团体客户", action = "修改")
    public ApiResponse<GroupCustomerLead> update(@PathVariable Long id, @RequestBody GroupCustomerLead body) {
        var target = repository.findById(id).orElseThrow();
        EntityPatchUtil.copyNonNull(body, target);
        return ApiResponse.ok(repository.save(target));
    }

    @DeleteMapping("/{id}")
    @OperationAudit(module = "团体客户", action = "删除")
    public ApiResponse<Void> delete(@PathVariable Long id) {
        repository.deleteById(id);
        return ApiResponse.ok();
    }
}