# obai_fullstack_package-v2/backend/src/main/java/com/obai/platform/controller/admin/AdminHealthArchiveController.java
package com.obai.platform.controller.admin;

import com.obai.platform.audit.OperationAudit;
import com.obai.platform.common.ApiResponse;
import com.obai.platform.common.EntityPatchUtil;
import com.obai.platform.entity.HealthArchive;
import com.obai.platform.repository.HealthArchiveRepository;
import java.util.List;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/admin/health-archives")
public class AdminHealthArchiveController {
    private final HealthArchiveRepository repository;

    public AdminHealthArchiveController(HealthArchiveRepository repository) {
        this.repository = repository;
    }

    @GetMapping
    public ApiResponse<List<HealthArchive>> list() {
        return ApiResponse.ok(repository.findAll());
    }

    @GetMapping("/{id}")
    public ApiResponse<HealthArchive> detail(@PathVariable Long id) {
        return ApiResponse.ok(repository.findById(id).orElseThrow());
    }

    @PostMapping
    @OperationAudit(module = "健康档案", action = "新增")
    public ApiResponse<HealthArchive> create(@RequestBody HealthArchive body) {
        return ApiResponse.ok(repository.save(body));
    }

    @PutMapping("/{id}")
    @OperationAudit(module = "健康档案", action = "修改")
    public ApiResponse<HealthArchive> update(@PathVariable Long id, @RequestBody HealthArchive body) {
        var target = repository.findById(id).orElseThrow();
        EntityPatchUtil.copyNonNull(body, target);
        return ApiResponse.ok(repository.save(target));
    }

    @DeleteMapping("/{id}")
    @OperationAudit(module = "健康档案", action = "删除")
    public ApiResponse<Void> delete(@PathVariable Long id) {
        repository.deleteById(id);
        return ApiResponse.ok();
    }
}